Add save session capability for tmux terminals
Tmux can be called automatically on login. The terminal should run
scripts/tmux-setup.sh to start a session properly. The script creates a
new session with the proper name (terminal or terminal-n). The script
also handles cleanup. On logout, if the tmux session is no longer
attached and hasn't been marked non-killable in ~/.tmux-saves/ the
script will kill the session.

To mark a tmux terminal session as non-killable, run bin/tmux-save
(aliased to tsave in aliases.sh). This creates a file with the name of
the tmux session in ~/.tmux-saves/.

The above only happen for tmux sessions that are started automatically
using tmux-setup.sh. Manually starting a tmux session will not have any
of this capability.

Aliases added:
- tkill: tmux kill-session
- tsave: tmux-save (~/bin/tmux-save)