Add tmux + vim integration
Add the vim-tmux-navigator plugin. This allows the use of <C-hjkl> to
switch between tmux splits as well as vim splits (or between the two of
them).

Also sets vi mode for selection in tmux which allows many vim
keybindings to be used (such as hjkl, wbe, HML, C-ud, $0...). Customized
to also include v for selection, C-a V for line selection, and V to
toggle block selection.